Extended dtype for real/ieee/64-bit float values
This patch extends the system for calculating 'dtype' to
deal with a wider range of 'nbits', including special logic
to handle 64-bit floating points.  In summary:

*) Writing a float64 array requires nbits=64, and writes out
an array with dtype=5 (E64)

*) Writing a float32 array with nbits=32 will write out E32
or e32 (5/133), as appropriate for array size; this was
formerly (1/134), but that didn't compress.

*) Writing a float32 array with 8 < nbits < 32 will write
out R# or f# (1/134) as appropriate.  This was the prior
default.

*) Reading .fst files with E64 values now properly allocates
a float64 array.

The upshot is that rpnstd will now allow reading/writing of
64-bit floats, and further output with nbits=32 will be
losslessly compressed now where it wasn't before.