Comptue datev upon reading a record
Datev is an important data member of an RPNRec, since it's
one of the records suitable for searching (it's part of
RPNKeys, and is used in RPNFile.info() which calls
fstinf/inl.  However, this isn't a key actually read via
fstprm, so a newly-read record will have a datev of -1.

This patch tweaks RPNFile.info so that when a valid dateo is
read (one that gives a proper date with RPNDate.toDateTime()
rather than a ValueError), the datev is calculated via dateo
+ deet*npas (seconds), cast back into an RPN Date Stamp
format.

Note for reference, the RPN Date Stamp format as a time
resolution of 5 seconds.