Update RPNFile.__setitem__ for noncontiguous values
For Fortran-style arrays, RPNFile.__setitem__ typically
passes the array to fstecr unmodified.  This is proper for
most arrays, but if instead the array being passed in is a
noncontiguous view of another array (as, for example,
data[:,::-1] to reverse a dimension), then improper values
were being assumed by fstecr; this would lead to accessing
out-of-bounds memory to write to the file.

This update checks the array flags for fortran-style arrays
to verify that the f_contiguous flag is set; if it is not
then 'value' is replaced at writeout with a
newly-instantiated, contiguous array.