Added implementation of ezscint wind-conversion
This update adds support to the Fstdc c-module for gduvfwd
and gdwdfuv, the routines that convert from grid-directed
(UU,VV) fields to (strength,direction) pairs.  An interface
is not yet built into RPNGrid for the conversion, but it
should be straightforward to implement.

Skeletal header information is included for a number of
other ezscint routines, but that implementation is still
pending.