Implement RPNGrid.getLatLon(x,y)
This patch implements RPNGrid.getLatLon(x,y), which acts as
a wrapper around gdllfxy.

If x and y are None (or not specified), the function returns
2D arrays corresponding to the (lat,lon) pairs at every
on-grid point.

If x and y are specified and have matching structures
(either both scalar or convertible to numpy.array with the
same shape), those values will be taken as coordinates for
the conversion; the output (lat,lon) pairs will be
numpy.arrays of the same shape as array(x).