Modify RPNRec.info to find RPNMeta if handle is nonzero
This patch modifies RPNRec.info() such that if it is passed
a bare RPNKeys instance (rather than a RPNMeta instance)
that happens to have a nonzero handle, a full search for a
record is still performed.  Previously the key was simply
returned, and this was an error.

This workflow shows up as:

foo = RPNFile(...)
uulist = foo.info(RPNKeys(nom='uu',list=True) # get all uu
for uumeta in uulist:
   # Find matching vv
   vvmeta = foo.info(RPNKeys(uumeta,nom='vv'))
   # Above line was returning the bare RPNKeys, because the
   # 'handle' data member was set to point at the location
   # of the UU record.
   uu = foo[uumeta] # Worked
   vv = foo[vvmeta] # Failed

This patch also fixes an out-of-bounds list reference (which
resulted in a runtime exception) when executing code of the
form:

> var_list = TheRPNFile.info(key,list=True)

if there are no matching variables in the file.  The error
occurred when setting the file's last-read-entry variable,
as there were no variables read.