Statically link intel libraries
The intel shared libraries were not being statically linked
in the build process.  This causes errors when importing
rpnstd on systems that do not have icc properly set up, from
the missing intel .so files.

The fix here is to explicitly include the LDFLAGS
environment variable at link-time in setup.py, combined with
specifying appropriate flags ("-static-intel") in the
compiler-specific makefile.