Restored buildability for PGI compilers
With the move to intel compilers as the primary build
environment, setup.py was invoked with the
--compiler=$(CCNAME) flag to use the intel compiler.  This
environment variable was not being specified (as 'unix') in
the pgi* makefiles; this is amended.

Additionally, LIBRMNSHARED was hard-coded in both setup.py
and the Makefile, while (in the current build environment)
it was only actually used in the former.  By reading the
environment variable, these uses have been linked: setup.py
will now use the value specified in LIBRMNSHARED to find the
appropriate rmnlib library.  This will also make moving to
non-release-candidate versions of said library easier in the
future.

(This also allows overriding LIBRMNSHARED on an
EC_ARCH-makefile basis, which will be convenient if
continuing support for older architectures is desirable).