Add compatibility extensions for numpy 1.3
Development of Fstdc wrappers for some of the interpolation
routines in rmnlib was done with a locally-installed version
of numpy 1.7.  This is substantially newer than the
globally-installed 1.3, and it has a slightly different API
in certain respects related to array flags and creation.

In order to retain compatibility with both older and newer
versions of numpy (presumably the globally-installed version
will be updated at some point), those missing extensions
were added as appropriately-implemented stubs in Fstdc.c,
with #ifdef blocks corresponding to numpy API versions.