Support for datyp=7 (string arrays) in .fst files
This changeset introduces support for character strings in
.fst files.  In the standard-format file, these strings have
datyp=7.

Specific changes include:

   *) Support in Fstdc.c and py_capi_ftn_utils.c for
datyp=7, creating and using a Numpy array with string
datatype when appropriate.

   *) Support for reading these arrays in rpnstd.py.  For
usability reasons, space-padded-to-max-length Fortran
strings are stripped of the space padding, giving directly-
usable but variable-length Python strings.  (The resulting
numpy array has a datatype of (np.str_,<maxlength>)).

      The dimension is also reduced, so that a record of
size (ni,nj) in the file, which corresponds to (nj) strings
of maximum length (ni), appears in Python as a
one-dimensional numpy array of length (nj).

   *) Workaround for a version incompatibility, where
in rmnlib < 15.1 string arrays were written with their high
bit flipped. This round-tripped fine within a single library
version, but caused reading gibberish data when a file was
read with a different version than used at its creation.
The workaround is a heuristic.

   *) Support for writing string arrays into .fst files with
datyp=7.  For compatibility with Fortran output, strings are
space-padded to their maximum length, which is the greater
of the <maxlen> in the numpy datatype or .ni of the
associated RPNMeta.