inp_open: add check
The HRDPS had a problem the other day where the pilot was running slowly
and didn't produce the input files for the nested grid in time (i.e. by
when the nested grid needed them).  The nested model died of course
(that's normal), but the dude on shift flagged the error message as
strange because it referred to MODEL_ANALYSIS rather than MODEL_INREP
even though the model was looking for LBCs.  Really not a big deal, but
they brought it up as a minor infrastructure thing.

In inp_open, we fall back from INREP to ANALYSIS if the 'content' file
is not found under INREP.  Then we fnom() the resulting path without
checking for existence first.  No real problem, since the model is about
to die with a decent "input not found" message anyways.  All that it
means is that we also get the fnom() error message that refers to
ANALYSIS.  The solution is pretty simple: check for the 'content' file
before trying to open it even in the fallback case, and then emit an
error message that contains the base path name for interested users if
no data is found.

This seems to be working alright in
/users/dor/armn/rmt/home/gem/5.0.b4/inpopen/inp_open.F90 in both the
case where the file is missing and where it's present.  And full
disclosure:  I wrote a goto because there were already a bunch of goto's
in there.  I know that I'll probably lose a gold star for that, but this
code is working fine so I didn't see the need to put in anything more
than minimal effort ;)