There is a potential div-by-zero in condload() if the square of the vertical motion is zero and buoyancy is small. The former can arise in the case of a very small computed vertical motion at the LCL: the WU at this level is 10^(-21) at the point that crashes, so that when squared WTW=0 via an underflow.
Because the test for a positive vertical motion is done on WU rather than WTW, the updraft calculations can proceed despite the 0 boundary condition for the integral.  In a statically unstable state, the updraft can accelerate even with a null boundary conditions, so a cloud could form.  This means that the answer is *not* to check for WTW>0.

We seem to get into this trouble through cloud objects.  With the decay imposed, it's not unreasonable to expect any physically realistic value of vertical motion, including those that are very close to zero.  It's not clear why this situation doesn't occur with the standard initiation trigger, but Paul is doing a bit of work to see what range of values we actually end up with out of interest.

Our real problem is that we're in RC3 and the final cycles are done for most systems.  We cannot change the bitpattern for cases that don't fail.  In future, some simple conditions will eliminate this problem, but for now we cannot see a way to implement them without potentially changing solutions.

We have confirmed that this approach does not change the bitpattern for runs that didn't crash, but prevents the crash from occurring in the case discussed here.  We are therefore the winz.

The code is available at:

/home/rmt001/home/gem/5.0.rc3/jingfix/for_rc4_2

There is lots of Intel-specific content in there, so we shouldn't forget to wrap those parts of the code in #ifdef blocks.