"Fix" the interface to c_mrfopr, so the 'MISSING' value can be set properly.
Note: This seems to fix the problem, but I have no idea why.  The source code
for c_mrfopr clearly uses a float for the value, so the ctypes argument
*should* be a c_float.  However, that always sends a value of 0 into c_mrfopr.

Setting the argument type to c_double seems to correctly pass the value to
c_mrfopr as a float across all ranges of values.  I don't know how or why.
I'm just going to chalk it up to a weird glitch in ctypes.